requirements.txt
dnspython==2.4.2
loguru==0.7.2
requests==2.31.0